Code quality improvements

+++ exchange/src/process_spawning/spawning.hpp
@@ -30,9 +30,9 @@ glz::json_t::object_t get_all_users();
  * @param users The ClientManager to spawn clients for
  * @returns the number of clients spawned
  */
-int spawn_all_clients(const nutc::manager::ClientManager& users);
+size_t spawn_all_clients(const nutc::manager::ClientManager& users);
 
-int initialize(manager::ClientManager& users, Mode mode);
+size_t initialize(manager::ClientManager& users, Mode mode);
 
 } // namespace client
 } // namespace nutc
